Remove test/client module
Move the integration test model into test/it and delete the old model
module. There is no need for this to be in a separate module - it will
only ever be required by the integration tests, and this had the
unfortunate effect of making the tests seem more complicated than they
actually are.